processor/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0